backend/requirements.txt
# Core Dependencies
pydantic>=2.10
pydantic-settings>=2.6

# Server
fastapi>=0.115
uvicorn[standard]>=0.32

# Database
sqlalchemy>=2.0
asyncpg>=0.30
alembic>=1.14
psycopg2-binary>=2.9

# Auth
python-jose[cryptography]>=3.4
bcrypt>=4.0
email-validator>=2.0

# Cache
redis>=5.0

# HTTP Client
httpx>=0.27

# Rate Limiting
slowapi>=0.1

# Testing
pytest>=8.3
pytest-asyncio>=0.24

# Utils
python-multipart>=0.0.9
python-dotenv>=1.0